feat: Embed new-org form
https://harperdb.atlassian.net/browse/STUDIO-480
