+++ src/features/organizations/index.tsx
@@ -1,13 +1,16 @@
 import { ConfirmDeletionModal } from '@/components/ConfirmDeletionModal';
 import { SubNavMenu } from '@/components/SubNavMenu';
+import { Button } from '@/components/ui/button';
 import { Input } from '@/components/ui/input';
 import { getCurrentUserQueryOptions } from '@/features/auth/queries/getCurrentUser';
 import { OrgCard } from '@/features/organizations/components/OrgCard';
-import { NewOrganizationModal } from '@/features/organizations/modals/NewOrganizationModal';
 import { useDeleteOrganizationMutation } from '@/features/organizations/mutations/deleteOrganization';
+import { NewOrg } from '@/features/organizations/NewOrg';
 import { curryFilterByFuzzySearch } from '@/lib/string/filterByFuzzySearch';
 import { queryKeys } from '@/react-query/constants';
 import { useQueryClient, useSuspenseQuery } from '@tanstack/react-query';
+import { Link } from '@tanstack/react-router';
+import { PlusIcon } from 'lucide-react';
 import { FormEvent, useCallback, useMemo, useState } from 'react';
 import { toast } from 'sonner';
 
@@ -63,13 +66,17 @@ export function OrganizationsIndex() {
 				});
 			}
 		},
-		[deleteOrg, queryClient, setIsDeleteOrgModalOpen]
+		[deleteOrg, queryClient, setIsDeleteOrgModalOpen],
 	);
 
 	const onDeleteOrgModal = useCallback((orgRole: { organizationId: string; organizationName?: string }) => {
 		setDeleteOrgInfo(orgRole);
 		setIsDeleteOrgModalOpen(true);
 	}, []);
+
+	if (!organizationRoles.length) {
+		return <NewOrg />;
+	}
 
 	return (
 		<>
@@ -80,7 +87,11 @@ export function OrganizationsIndex() {
 						className="inline-block w-48 md:w-64 bg-black border"
 						onChange={onFilterByNameChanged}
 					/>
-					<NewOrganizationModal />
+					<Link to="/new-org">
+						<Button variant="positive" className="rounded-full w-44" accessKey="n">
+							<PlusIcon /> <span><u>N</u>ew Organization</span>
+						</Button>
+					</Link>
 				</div>
 			</SubNavMenu>
 			<section className="mt-40 md:mt-32 px-4 pt-4 md:px-12 min-h-[calc(100vh-theme(spacing.32))]">
