+++ src/features/organizations/routes.ts
@@ -1,6 +1,7 @@
-import { createRoute } from '@tanstack/react-router';
-import { dashboardLayout } from '@/router/dashboardRoute';
 import { OrganizationsIndex } from '@/features/organizations/index';
+import { NewOrg } from '@/features/organizations/NewOrg';
+import { dashboardLayout } from '@/router/dashboardRoute';
+import { createRoute } from '@tanstack/react-router';
 
 export const orgsLayoutRoute = createRoute({
 	getParentRoute: () => dashboardLayout,
@@ -13,4 +14,13 @@ const orgsIndexRoute = createRoute({
 	component: OrganizationsIndex,
 });
 
-export const orgsRoutes = [orgsIndexRoute];
+const newOrgRoute = createRoute({
+	getParentRoute: () => orgsLayoutRoute,
+	path: '/new-org',
+	component: NewOrg,
+});
+
+export const orgsRoutes = [
+	orgsIndexRoute,
+	newOrgRoute,
+];
